Refactor: like domain name을 likes로 변경
