+++ src/main/java/com/encore/playground/domain/likes/entity/Likes.java
@@ -1,4 +1,4 @@
-package com.encore.playground.domain.like.entity;
+package com.encore.playground.domain.likes.entity;
 
 import com.encore.playground.domain.feed.entity.Feed;
 import com.encore.playground.domain.member.entity.Member;
